+++ api/entity/transaction.go
@@ -3,7 +3,7 @@ package entity
 import "time"
 
 type Transaction struct {
-	IdTransaction uint64
+	IdTransaction int
 	ProductName   string
 	CustomerName  string
 	NoTransaction string
