+++ api/model/transaction_model.go
@@ -20,7 +20,7 @@ type GetProductNameTransactionResponse struct {
 }
 
 type GetTransactionResponse struct {
-	IdTransaction uint64 `json:"id_transaction"`
+	IdTransaction int    `json:"id_transaction"`
 	ProductName   string `json:"product_name"`
 	CustomerName  string `json:"customer_name"`
 	NoTransaction string `json:"no_transaction"`
